requirement.txt
# Core dependencies for Blackjack RL Agent
pygame==2.5.2
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
collections-extended==2.0.2

# Optional dependencies for enhanced analysis
jupyter==1.0.0
notebook==7.0.2
ipykernel==6.25.0

# Development and utility dependencies
python-dateutil>=2.8.2
pillow>=10.0.0
six>=1.16.0